KDE: migrate settings to Plasma 6
Now khotkeys have been removed and replaced by kglobalaccel. To get our
old global shortcuts back we had to create .desktop files for each
command.

And to keep the trouble of keeping the kglobalshortcutsrc up-to-date via
`kde-setup`, we check in the entire file.

Also change the configured theme to Reactionary/Windows/SE98.