Rewrite setup script to allow separate OS configurations
Rewrite the script in Python because shell scripting is too much of a
headache. Can expect Python to be available anyway.

Point of the rewrite though is to enable separate configuration for
Linux and Mac (and other unames in the futures, if so wanted).

Also no longer required to list each file to be included in the
symlinking, if it is available in the expected directory hierarchy it
will be symlinked.

Renamed install/uninstall to link/unlink to more clearly signify their
operation.

Dropped the 'install-icons' subcommand, have not used it in a long time
anyway but if wanted in the future, should have a more general way of
fetching and unpacking external assets.